+++ src/components/favorites-empty/favorites-empty.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import {Provider} from 'react-redux';
 import {BrowserRouter} from 'react-router-dom';
-import mockStore from '../../mocks/mock-store';
+import mockStore from '../../test-data/mock-store';
 
 import FavoritesEmpty from './favorites-empty.jsx';
