+++ src/components/favorites/favorites.test.js
@@ -4,8 +4,8 @@ import {Provider} from 'react-redux';
 import {BrowserRouter} from 'react-router-dom';
 
 import {Favorites} from './favorites.jsx';
-import mockStore from '../../mocks/mock-store.js';
-import offersMock from '../../mocks/offers.js';
+import mockStore from '../../test-data/mock-store.js';
+import offersMock from '../../test-data/offers.js';
 
 it(`Favorites renders correctly`, () => {
   const tree = renderer
