+++ src/components/login/login.test.js
@@ -3,7 +3,7 @@ import renderer from 'react-test-renderer';
 import {Provider} from 'react-redux';
 import {BrowserRouter} from "react-router-dom";
 import Login from './login.jsx';
-import mockStore from '../../mocks/mock-store.js';
+import mockStore from '../../test-data/mock-store.js';
 
 it(`Header renders correctly`, () => {
   const tree = renderer
