+++ src/components/main/main.test.js
@@ -3,8 +3,8 @@ import renderer from 'react-test-renderer';
 import {Provider} from 'react-redux';
 import Main from './main.jsx';
 import {BrowserRouter} from 'react-router-dom';
-import mockStore from '../../mocks/mock-store.js';
-import mockOffers from '../../mocks/offers.js';
+import mockStore from '../../test-data/mock-store.js';
+import mockOffers from '../../test-data/offers.js';
 
 
 it(`Main renders correctly`, () => {
