+++ src/components/offer-details/offer-details.test.js
@@ -4,8 +4,8 @@ import {Provider} from 'react-redux';
 import {OfferDetails} from './offer-details.jsx';
 import {BrowserRouter} from 'react-router-dom';
 
-import mockStore from '../../mocks/mock-store.js';
-import mockOffers from '../../mocks/offers.js';
+import mockStore from '../../test-data/mock-store.js';
+import mockOffers from '../../test-data/offers.js';
 
 const mockReviews = [
   {
