+++ src/hocs/with-map/with-map.test.js
@@ -3,8 +3,8 @@ import renderer from 'react-test-renderer';
 import PropTypes from 'prop-types';
 import {Provider} from 'react-redux';
 import withMap from './with-map.jsx';
-import mockStore from '../../mocks/mock-store.js';
-import mockOffers from '../../mocks/offers.js';
+import mockStore from '../../test-data/mock-store.js';
+import mockOffers from '../../test-data/offers.js';
 
 const MockComponent = ({mapRef}) => <div ref={mapRef}/>;
 const WrappedComponent = withMap(MockComponent);
