+++ src/hocs/with-sort/with-sort.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import {Provider} from 'react-redux';
 import withSort from './with-sort.jsx';
-import mockStore from '../../mocks/mock-store.js';
+import mockStore from '../../test-data/mock-store.js';
 
 const MockComponent = () => <div />;
 const WrappedComponent = withSort(MockComponent);
