+++ src/test-data/details.js
